requirements.txt
torch
torchvision
facenet-pytorch
pillow
opencv-python
pytesseract
streamlit
numpy
